+++ app/(auth)/dashboard/views/candidate/resume-copilot/builder.tsx
@@ -20,7 +20,7 @@ import {
 import { arrayMove, sortableKeyboardCoordinates } from "@dnd-kit/sortable";
 import Container from "./container";
 import SortableItem from "./sortable";
-import { Item, ItemType, CustomItem, ResumeBuilderProps } from "./types";
+import { Item, ItemType, CustomItem, ResumeBuilderProps, QueueItem } from "./types";
 import { buildEditReview } from "./prompt-builder"; // AI Call
 import { suggestNextSteps } from "./next-steps"; // AI Call #2
 import { Button } from "@/components/ui/button";
@@ -55,6 +55,7 @@ import {
 import ProcessingIndicator from "./processing-indicator";
 import CopilotTalk from "./copilot-talk";
 import ControlPanel from "./control-panel";
+import { handleDragEnd } from "./handler-drag-end";
 
 interface BuilderSession {
   sessionId: string;
@@ -65,11 +66,6 @@ interface CustomSection {
   id: string;
   title: string;
   items: CustomItem[];
-}
-
-interface QueueItem {
-  itemId: string;
-  cardContent: any;
 }
 
 interface Alert {
@@ -573,238 +569,47 @@ const ResumeBuilder: React.FC<ResumeBuilderProps> = ({
     [findContainer]
   );
 
-  const handleDragEnd = useCallback(
+  const dragEnd = useCallback(
     (event: DragEndEvent) => {
-      const { active, over } = event;
-      const id = active.id as string;
-      const overId = over?.id as string;
-
-      if (id === "custom-card" && overId) {
-        const targetSection = customSections.find((section) =>
-          section.items.some((item) => item.id === overId)
-        );
-        if (targetSection) {
-          handleAddCustomItem(targetSection.id);
-        } else {
-          const newSectionId = `custom-section-${Date.now()}`;
-          const newSection: CustomSection = {
-            id: newSectionId,
-            title: "New Custom Section",
-            items: [],
-          };
-          setCustomSections([...customSections, newSection]);
-          handleAddCustomItem(newSectionId);
-        }
-        return;
-      }
-
-      const sourceSection = customSections.find((section) =>
-        section.items.some((item) => item.id === id)
+      handleDragEnd(
+        event,
+        items,
+        customSections,
+        findContainer,
+        setItems,
+        setCustomSections,
+        setDragOrigin,
+        setActiveId,
+        setDraggedItem,
+        dragStartContainer,
+        excludedPersonalItems,
+        setLastModifiedItemId,
+        setActionHistory,
+        setProcessingQueue,
+        setProcessingItems,
+        handleAddCustomItem
       );
-      const targetSection = customSections.find((section) =>
-        section.items.some((item) => item.id === overId)
-      );
-
-      if (sourceSection && targetSection) {
-        setCustomSections((prevSections) =>
-          prevSections.map((section) => {
-            if (
-              section.id === sourceSection.id &&
-              section.id === targetSection.id
-            ) {
-              const oldIndex = section.items.findIndex(
-                (item) => item.id === id
-              );
-              const newIndex = section.items.findIndex(
-                (item) => item.id === overId
-              );
-              const newItems = arrayMove(section.items, oldIndex, newIndex);
-              return { ...section, items: newItems };
-            } else if (section.id === sourceSection.id) {
-              return {
-                ...section,
-                items: section.items.filter((item) => item.id !== id),
-              };
-            } else if (section.id === targetSection.id) {
-              const itemToMove = sourceSection.items.find(
-                (item) => item.id === id
-              )!;
-              const overIndex = section.items.findIndex(
-                (item) => item.id === overId
-              );
-              const newItems = [...section.items];
-              newItems.splice(overIndex, 0, itemToMove);
-              return { ...section, items: newItems };
-            }
-            return section;
-          })
-        );
-        return;
-      }
-
-      const activeContainer = findContainer(id);
-      const overContainer = findContainer(overId);
-
-      setItems((prevItems) => {
-        const activeItems =
-          prevItems[activeContainer as keyof typeof prevItems];
-        const overItems = prevItems[overContainer as keyof typeof prevItems];
-
-        if (!activeItems || !overItems) {
-          console.error("Invalid containers:", {
-            activeContainer,
-            overContainer,
-            prevItems,
-          });
-          return prevItems;
-        }
-
-        const activeIndex = activeItems.findIndex((item) => item.id === id);
-        const overIndex = overItems.findIndex((item) => item.id === overId);
-
-        let newItems = { ...prevItems };
-
-        if (activeContainer !== overContainer) {
-          newItems = {
-            ...newItems,
-            [activeContainer as keyof typeof prevItems]: activeItems.filter(
-              (item) => item.id !== id
-            ),
-            [overContainer as keyof typeof prevItems]: [
-              ...overItems.slice(0, overIndex),
-              activeItems[activeIndex],
-              ...overItems.slice(overIndex),
-            ],
-          };
-        } else {
-          newItems = {
-            ...newItems,
-            [overContainer as keyof typeof prevItems]: arrayMove(
-              overItems,
-              activeIndex,
-              overIndex
-            ),
-          };
-        }
-
-        if (
-          ((activeItems[activeIndex].type === "experience" ||
-            activeItems[activeIndex].type === "education") &&
-            overContainer === "chosen") ||
-          (overContainer === "chosen" &&
-            newItems["chosen"].some(
-              (item) => item.type === "experience" || item.type === "education"
-            ))
-        ) {
-          const allItems = newItems["chosen"];
-          const experienceAndEducationItems = allItems.filter(
-            (item) => item.type === "experience" || item.type === "education"
-          );
-          const otherItems = allItems.filter(
-            (item) => item.type !== "experience" && item.type !== "education"
-          );
-
-          const sortedExperienceAndEducationItems =
-            experienceAndEducationItems.sort((a, b) => {
-              if (
-                (a.type !== "experience" && a.type !== "education") ||
-                (b.type !== "experience" && b.type !== "education")
-              )
-                return 0;
-
-              const aEndDate = a.content.end_date.toLowerCase();
-              const bEndDate = b.content.end_date.toLowerCase();
-
-              if (aEndDate === "present" && bEndDate === "present") {
-                return b.content.start_date.localeCompare(a.content.start_date);
-              }
-
-              if (aEndDate === "present") return -1;
-              if (bEndDate === "present") return 1;
-
-              const endDateComparison = bEndDate.localeCompare(aEndDate);
-              if (endDateComparison !== 0) return endDateComparison;
-
-              return b.content.start_date.localeCompare(a.content.start_date);
-            });
-
-          const sortedExperienceItems =
-            sortedExperienceAndEducationItems.filter(
-              (item) => item.type === "experience"
-            );
-          const sortedEducationItems = sortedExperienceAndEducationItems.filter(
-            (item) => item.type === "education"
-          );
-
-          newItems["chosen"] = [
-            ...sortedExperienceItems,
-            ...sortedEducationItems,
-            ...otherItems,
-          ];
-        }
-
-        const movedItem = activeItems[activeIndex];
-        const isExcludedPersonalItem =
-          movedItem.type === "personal" &&
-          excludedPersonalItems.includes(movedItem.content.key);
-
-        if (!isExcludedPersonalItem) {
-          if (
-            dragStartContainer === "available" &&
-            overContainer === "chosen"
-          ) {
-            setLastModifiedItemId(id);
-
-            const newAction = {
-              action: "add",
-              itemId: id,
-              itemType: movedItem.type,
-              fromContainer: dragStartContainer,
-              toContainer: overContainer,
-              newIndex: overIndex,
-            };
-
-            setActionHistory((prevHistory) => [
-              ...prevHistory,
-              {
-                ...newAction,
-                action: newAction.action as "add" | "remove",
-                itemType: newAction.itemType as string,
-                toContainer: newAction.toContainer as string | null,
-              },
-            ]);
-
-            // Only add to processing queue and set as processing if it's not a personal item
-            if (
-              movedItem.type !== "personal" ||
-              (movedItem.type === "personal" &&
-                movedItem.content.key === "intro")
-            ) {
-              setProcessingQueue((prevQueue) => [
-                ...prevQueue,
-                { itemId: id, cardContent: movedItem.content },
-              ]);
-
-              setProcessingItems((prev) => new Set(prev).add(id));
-            }
-          }
-        }
-
-        return newItems;
-      });
-
-      setDragOrigin(null);
-      setActiveId(null);
-      setDraggedItem(null);
     },
     [
+      items,
       customSections,
       findContainer,
+      setItems,
+      setCustomSections,
+      setDragOrigin,
+      setActiveId,
+      setDraggedItem,
       dragStartContainer,
       excludedPersonalItems,
-      handleAddCustomItem,
+      setLastModifiedItemId,
+      setActionHistory,
+      setProcessingQueue,
+      setProcessingItems,
+      handleAddCustomItem
     ]
   );
+  
+  
 
   /* Editor */
   const handleEdit = useCallback((item: Item) => {
@@ -1234,7 +1039,7 @@ const ResumeBuilder: React.FC<ResumeBuilderProps> = ({
       collisionDetection={closestCorners}
       onDragStart={handleDragStart}
       onDragOver={handleDragOver}
-      onDragEnd={handleDragEnd}
+      onDragEnd={dragEnd}
     >
       <div className="flex gap-6 h-[calc(100vh-12rem)] overflow-hidden">
         {renderAvailableItems}
